refactor and move DoiField
